requirements.txt
aiohttp>=3.9.0
aiohttp-jinja2>=1.6
pyyaml>=6.0
monobank>=0.4.0
